requirements.txt
pandas
numpy
scikit-learn
yfinance
transformers
torch
peft
accelerate
mlflow
pytest
pyyaml
pydantic
tqdm
rich
matplotlib
